# services/internal/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Chatbook Microservices
    version: "1.0"
servers:
    - url: http://localhost:80
      description: Default server for chatbook-microservices
paths:
    /api/v1/auth/login:
        post:
            tags:
                - auth
            summary: login auth
            operationId: auth#login
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequestBody'
                        example:
                            email: Provident harum similique saepe fuga assumenda.
                            password: Sequi libero et qui soluta error.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
                            example:
                                email: Et delectus qui.
                                id: Quas voluptas ducimus perferendis assumenda eum officiis.
                                refresh_token: Sed quia eveniet sit.
                                role: Ut blanditiis qui.
                                token: Veniam tempore voluptate sunt.
                "404":
                    description: 'WrongCredentials: Email and/or password don''t match'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
    /api/v1/auth/refresh-token:
        post:
            tags:
                - auth
            summary: refreshToken auth
            operationId: auth#refreshToken
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RefreshTokenRequestBody'
                        example:
                            email: Assumenda autem quo aut.
                            refreshToken: Reprehenderit totam.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
                            example:
                                email: Quibusdam voluptatum voluptatum vitae.
                                id: Sit veniam delectus.
                                refresh_token: Qui vel voluptas aperiam quae.
                                role: Explicabo omnis nulla.
                                token: Qui rerum quidem eum id eveniet.
...
    /api/v1/users/profile:
        put:
            tags:
                - user
            summary: updateProfileNames user
            operationId: user#updateProfileNames
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateProfileNamesRequestBody'
                        example:
                            firstName: Quibusdam qui.
                            lastName: Tempora perspiciatis ut ut.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OperationStatusResponse'
                            example:
                                message: Optio a.
...
components:
    schemas:
        Error:
            type: object
            properties:
                fault:
                    type: boolean
                    description: Is the error a server-side fault?
                    example: true
                id:
                    type: string
                    description: ID is a unique identifier for this particular occurrence of the problem.
                    example: 123abc
                message:
                    type: string
                    description: Message is a human-readable explanation specific to this occurrence of the problem.
                    example: parameter 'p' must be an integer
                name:
                    type: string
                    description: Name is the name of this class of errors.
                    example: bad_request
                temporary:
                    type: boolean
                    description: Is the error temporary?
                    example: true
                timeout:
                    type: boolean
                    description: Is the error a timeout?
                    example: false
            example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
            required:
                - name
                - id
                - message
                - temporary
                - timeout
                - fault
...
        LoginRequestBody:
            type: object
            properties:
                email:
                    type: string
                    description: User email
                    example: Quasi eum dignissimos excepturi.
                password:
                    type: string
                    description: User password
                    example: Ut harum.
            example:
                email: Quaerat iure quod ut voluptate.
                password: Quis eligendi ullam quasi enim laudantium aliquam.
            required:
                - email
                - password
        LoginResponse:
            type: object
            properties:
                email:
                    type: string
                    description: User's email
                    example: Rerum soluta ut deleniti esse.
                id:
                    type: string
                    description: User account's ID
                    example: Error veritatis.
                refresh_token:
                    type: string
                    description: Refresh token
                    example: Magnam totam dolorem.
                role:
                    type: string
                    description: User's role
                    example: Eaque vel.
                token:
                    type: string
                    description: JWT Token
                    example: Quod at aspernatur et repellendus earum.
            example:
                email: Doloremque maxime molestiae natus voluptatem.
                id: Aliquam beatae ipsam aut.
                refresh_token: Qui quasi rerum fugit cum.
                role: Ut doloremque qui voluptatem sed odit.
                token: Omnis ut quod quia beatae magni consequatur.
            required:
                - email
                - token
                - refresh_token
                - role
        OperationStatusResponse:
            type: object
            properties:
                message:
                    type: string
                    description: Operation status
                    example: Optio error est ut velit.
            example:
                message: Numquam molestiae dolorem.
            required:
                - message
        RefreshTokenRequestBody:
            type: object
            properties:
                email:
                    type: string
                    description: User email
                    example: Et ratione perspiciatis.
                refreshToken:
                    type: string
                    description: Refresh token
                    example: Quia sapiente qui.
            example:
                email: Aut quia minus dolores.
                refreshToken: Excepturi dolore blanditiis ut.
            required:
                - email
                - refreshToken
...
        UpdateProfileNamesRequestBody:
            type: object
            properties:
                firstName:
                    type: string
                    description: Updated user first name
                    example: Quia ab est repellendus exercitationem aut delectus.
                lastName:
                    type: string
                    description: Updated user last name
                    example: Ex repellendus.
            example:
                firstName: Sapiente nostrum minima reiciendis.
                lastName: Qui beatae deserunt nihil qui recusandae hic.
            required:
                - firstName
                - lastName
...